bot/requirements.txt
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
jalali_core==1.0.0
jdatetime==5.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pyTelegramBotAPI==4.14.0
python-dotenv==1.1.1
pytz==2025.2
requests==2.31.0
schedule==1.2.0
urllib3==2.5.0
Werkzeug==3.1.3
qrcode[pil]
